When a domain query times out, create a new instance and try again
Z3 seems to be a bit unstable with integer problems, so this prevents the solver from entering an endless loop by adding a timeout to domain queries.
If the timeout is reached, the domain instance is rebuilt (unless it has just been rebuilt) and the query is tried again.
If it fails yet again, the domain throws an error.