Rewrote benchmark utility:
It runs vvt on all benchmarks in a given directory.
*Logs for all runs are written to ./log/ or to a specified
directory.
*Stats for all benchmarks are collected in a file, with mean
values for all fields in IC3Stats printed in the end.
*Transitionrelations can be cached (default dir: ./trcache/ ),
the program will not build the transition relation if it finds
one here with the same name.

Added possibilty to dump states that were collected for refinement
into a file in csv format. (see vvt-verify option --dump-states-to )

Added possibilty to dump stats of a verify run into a machine-readable
format (json). See vvt-verify option --dump-stats-to.